fix: 人力対戦ゲームでのゲーム表示を調整
