+++ pages/game/manual/index.tsx
@@ -13,19 +13,14 @@ import {
   apiClient,
   Game,
   WsGameReq,
-  WsGameRes,
   MatchRes,
   ActionPost,
-  host,
 } from "../../../src/apiClient";
-import Link from "../../../src/link";
 import { useGameUsers } from "../../../src/useGameUsers";
-import datas from "../../../components/player_datas";
 
+import datas from "../../../components/player_datas";
+import GamePanel from "../../../components/gamePanel";
 import Content from "../../../components/content";
-import GameList from "../../../components/gamelist";
-import GameBoard from "../../../components/gameBoard";
-import PointsGraph from "../../../components/pointsGraph";
 
 type NextActionType = [-1 | 0 | 1, -1 | 0 | 1];
 const NextActions: Record<
@@ -249,42 +244,15 @@ const Page: NextPage<{ id?: string }> = ({ id }) => {
     };
   }, [updateGamepads]);
 
-  const connect = useCallback(() => {
-    if (!matchRes) return;
-    const socket = new WebSocket(
-      (host.protocol === "https:" ? "wss://" : "ws://") +
-        host.host +
-        "/v1/ws/game"
-    );
-    socket.onopen = () => {
-      const q = `id:${matchRes.gameId}`;
-      console.log(q);
-      const req: WsGameReq = {
-        q,
-      };
-      socket.send(JSON.stringify(req));
+  const query = useMemo(() => {
+    if (!matchRes) return undefined;
+    const q = `id:${matchRes.gameId}`;
+    console.log(q);
+    const req: WsGameReq = {
+      q,
     };
-    socket.onmessage = (event) => {
-      const res = JSON.parse(event.data) as WsGameRes;
-      console.log(res);
-      let game: Game | undefined;
-      if (res.type === "initial") {
-        game = res.games[0];
-      } else {
-        game = res.game;
-      }
-      // console.log("setGame");
-      setGame(game);
-    };
-    return () => {
-      socket.close();
-      console.log("websocket close");
-    };
+    return req;
   }, [matchRes]);
-
-  useEffect(() => {
-    return connect();
-  }, [connect]);
 
   useEffect(() => {
     setControllerList((prev) => {
@@ -565,18 +533,15 @@ const Page: NextPage<{ id?: string }> = ({ id }) => {
             );
           })}
         </Box>
-        {game && (
-          <>
-            <Button
-              href={id ? `/vr/index.html?id=${id}` : "/vr/latest.html"}
-              style={{ margin: "auto" }}
-            >
-              VR版はこちら
-            </Button>
-            <GameList games={[game]} pagenation={false} hover={false} />
-            <GameBoard users={users} game={game} />
-            <PointsGraph users={users} game={game} />
-          </>
+        {query && (
+          <Box
+            sx={{
+              height: "calc(2em + 50vw)",
+              maxHeight: "calc(100vh - 64px)",
+            }}
+          >
+            <GamePanel query={query} />
+          </Box>
         )}
       </div>
     </Content>
